+++ h2integrate/core/supported_models.py
@@ -12,10 +12,6 @@ from h2integrate.storage.hydrogen.tank_baseclass import (
     HydrogenTankPerformanceModel,
 )
 from h2integrate.converters.wind.wind_plant_pysam import PYSAMWindPlantPerformanceModel
-from h2integrate.converters.ammonia.ammonia_baseclass import (
-    AmmoniaCostModel,
-    AmmoniaPerformanceModel,
-)
 from h2integrate.converters.desalination.desalination import (
     ReverseOsmosisCostModel,
     ReverseOsmosisPerformanceModel,
@@ -30,6 +26,10 @@ from h2integrate.converters.methanol.smr_methanol_plant import (
     SMRMethanolPlantFinanceModel,
     SMRMethanolPlantPerformanceModel,
 )
+from h2integrate.converters.ammonia.simple_ammonia_model import (
+    SimpleAmmoniaCostModel,
+    SimpleAmmoniaPerformanceModel,
+)
 from h2integrate.converters.methanol.co2h_methanol_plant import (
     CO2HMethanolPlantCostModel,
     CO2HMethanolPlantFinanceModel,
@@ -42,6 +42,16 @@ from h2integrate.converters.water.hydro_plant_run_of_river import (
 )
 from h2integrate.converters.hydrogen.eco_tools_pem_electrolyzer import (
     ECOElectrolyzerPerformanceModel,
+)
+from h2integrate.converters.hydrogen.geologic.natural_geoh2_plant import (
+    NaturalGeoH2CostModel,
+    NaturalGeoH2FinanceModel,
+    NaturalGeoH2PerformanceModel,
+)
+from h2integrate.converters.hydrogen.geologic.stimulated_geoh2_plant import (
+    StimulatedGeoH2CostModel,
+    StimulatedGeoH2FinanceModel,
+    StimulatedGeoH2PerformanceModel,
 )
 
 
@@ -64,8 +74,8 @@ supported_models = {
     "hopp": HOPPComponent,
     "reverse_osmosis_desalination_performance": ReverseOsmosisPerformanceModel,
     "reverse_osmosis_desalination_cost": ReverseOsmosisCostModel,
-    "ammonia_performance": AmmoniaPerformanceModel,
-    "ammonia_cost": AmmoniaCostModel,
+    "simple_ammonia_performance": SimpleAmmoniaPerformanceModel,
+    "simple_ammonia_cost": SimpleAmmoniaCostModel,
     "steel_performance": SteelPerformanceModel,
     "steel_cost": SteelCostAndFinancialModel,
     "smr_methanol_plant_performance": SMRMethanolPlantPerformanceModel,
@@ -74,6 +84,12 @@ supported_models = {
     "co2h_methanol_plant_performance": CO2HMethanolPlantPerformanceModel,
     "co2h_methanol_plant_cost": CO2HMethanolPlantCostModel,
     "co2h_methanol_plant_financial": CO2HMethanolPlantFinanceModel,
+    "natural_geoh2_performance": NaturalGeoH2PerformanceModel,
+    "natural_geoh2_cost": NaturalGeoH2CostModel,
+    "natural_geoh2": NaturalGeoH2FinanceModel,
+    "stimulated_geoh2_performance": StimulatedGeoH2PerformanceModel,
+    "stimulated_geoh2_cost": StimulatedGeoH2CostModel,
+    "stimulated_geoh2": StimulatedGeoH2FinanceModel,
     # Transport
     "cable": CablePerformanceModel,
     "pipe": PipePerformanceModel,